SCC-2235 Add susbsidiary fields for Call Number data
To represent data in the new Items table in SCC there need to be two new fields that represent the fields that make up the `shelfMark` in separate fields. The first is `bf:physicalLocation` which represents the root fields of the `shelfMark`, the second is `bf:enumerationAndChronology` which represents the `v` fieldTag suffix data. This does not affect the original `shelfMark` field as it is purely additive, so this does not anticipate any side effects.