Fix locations.json, improve validate-locations.py
This rebuilds json-ld/locations.json. The version in `scc-2738` was
missing `skos:notation` elements for some reason.

I believe the absense of those keys caused the `validate-locations.py`
script to fail, which caused it to fail to switch *back* to `master`
branch, which caused subsequent invocations to report 0 changes between
locations.json versions (because it was comparing the version in
`master` to itself. This branch includes small fixes to that script to
ensure it switches back to the original branch regardless of what
exceptions are thrown. Also adds a check to report on missing
`skos:notation` keys when discovered.