Reconfigure how parallels are related in mapping
Trying a different way of relating parallel fields to their counterpart
in field-mapping-bib.json. Rather than repeat all marc paths in the
parallel mapping, let's just add a "isParallelFor" at the mapping level
(rather than path level) to 1) ensure all of the marc paths used for the
primary mapping are also used for the parallel mapping, and 2) allow us,
when building out the parallel values array, to ensure the indexes of
parallel values match up exactly with the primary values.